[flang][OpenMP] Privatize "loop-local" values in `do concurent` on device
Extends https://github.com/ROCm/llvm-project/pull/112.

This PR extends support for `do concurrent` mapping to the device a bit
more. In particular, it handles localization of loop-local values on the
deive. Previously, this was only supported and tested on the host.

See docs for `looputils::collectLoopLocalValues` for the definition of
"loop-local" values.